backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Data validation
pydantic==2.7.1
pydantic-settings==2.3.0

# FREE embeddings — runs fully locally, no API key needed (~90 MB model download)
sentence-transformers==3.0.1
# CPU-only torch (smaller image, no GPU needed)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1

# FREE LLM — Groq REST API (httpx client, no openai package needed)
httpx==0.27.0

# Vector DB — Pinecone free starter tier
pinecone-client==4.1.0

# Cache
redis==5.0.7

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
loguru==0.7.2
numpy==1.26.4

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
